fix(#8311): does this help?

+++ src/panel/JournalList.cpp
@@ -1235,7 +1235,7 @@ void JournalList::setAutomaticColumnSize()
     // calculate and apply diff:
     int diff = this->GetSize().GetWidth() - twidth;
     if (abs(diff) > 5) {
-        int diffdelta = diff / resizable_ids.size();
+        int diffdelta = diff / static_cast<int>(resizable_ids.size());
         for (colInfo col: resizable_ids) {
             this->SetColumnWidth(col.id, col.width + diffdelta);
         }
